ML integration/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
tensorflow==2.17.1
numpy==1.26.4
Pillow==10.2.0
